requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.6,!=0.7,<1.0
Babel>=1.3

oslo.concurrency>=0.3.0  # Apache-2.0
oslo.config>=1.6.0
oslo.context>=0.1.0  # Apache-2.0
oslo.db>=1.3.0  # Apache-2.0
oslo.messaging>=1.4.0,!=1.5.0
oslo.serialization>=1.2.0               # Apache-2.0
oslo.utils>=1.2.0                       # Apache-2.0
paramiko>=1.13.0
pecan>=0.8.0
keystonemiddleware>=1.0.0
python-heatclient>=0.2.9
python-keystoneclient>=0.11.1
python-zaqarclient>=0.0.3
six>=1.7.0
SQLAlchemy>=0.9.7,<=0.9.99
WSME>=0.6
jsonpatch>=1.1

